move all interfaces into components package
