+++ sia/hostdb/hostannouncement.go
@@ -3,34 +3,13 @@ package hostdb
 import (
 	"github.com/NebulousLabs/Sia/consensus"
 	"github.com/NebulousLabs/Sia/encoding"
-	"github.com/NebulousLabs/Sia/network"
+	"github.com/NebulousLabs/Sia/sia/components"
 )
 
-const (
-	HostAnnouncementPrefix = "HostAnnouncement"
-)
-
-// A HostAnnouncement is a struct that can appear in the arbitrary data field.
-// It is preceded by 8 bytes that decode to the integer 1.
-type HostAnnouncement struct {
-	IPAddress          network.Address
-	TotalStorage       int64 // Can go negative.
-	MinFilesize        uint64
-	MaxFilesize        uint64
-	MinDuration        consensus.BlockHeight
-	MaxDuration        consensus.BlockHeight
-	MinChallengeWindow consensus.BlockHeight
-	MaxChallengeWindow consensus.BlockHeight
-	MinTolerance       uint64
-	Price              consensus.Currency
-	Burn               consensus.Currency
-	CoinAddress        consensus.CoinAddress // Host may want to give different addresses to each client.
-
-	SpendConditions consensus.SpendConditions
-	FreezeIndex     uint64 // The index of the output that froze coins.
-}
-
-func findHostAnnouncements(height consensus.BlockHeight, b consensus.Block) (entries []HostEntry, err error) {
+// findHostAnnouncements scans a block and pulls out every host announcement
+// that appears in the block, returning a list of entries that correspond with
+// the announcements.
+func findHostAnnouncements(height consensus.BlockHeight, b consensus.Block) (entries []components.HostEntry, err error) {
 	for _, t := range b.Transactions {
 		// Check the arbitrary data of the transaction to fill out the host database.
 		if len(t.ArbitraryData) == 0 {
@@ -42,7 +21,7 @@ func findHostAnnouncements(height consensus.BlockHeight, b consensus.Block) (ent
 
 		dataIndicator := encoding.DecUint64([]byte(t.ArbitraryData[0][0:8]))
 		if dataIndicator == 1 {
-			var ha HostAnnouncement
+			var ha components.HostAnnouncement
 			err = encoding.Unmarshal([]byte(t.ArbitraryData[0][8:]), &ha)
 			if err != nil {
 				return
@@ -66,7 +45,7 @@ func findHostAnnouncements(height consensus.BlockHeight, b consensus.Block) (ent
 
 			// Add the host to the host database.
 			entryID := t.OutputID(int(ha.FreezeIndex))
-			entries = append(entries, HostEntry{
+			entries = append(entries, components.HostEntry{
 				ID:          string(entryID[:]),
 				IPAddress:   ha.IPAddress,
 				MinFilesize: ha.MinFilesize,
