+++ sia/hostdb/hostdb.go
@@ -1,36 +1,133 @@
 package hostdb
 
 import (
+	"crypto/rand"
+	"errors"
+	"math/big"
+	"sync"
+
 	"github.com/NebulousLabs/Sia/consensus"
+	"github.com/NebulousLabs/Sia/sia/components"
 )
 
-// The HostDB interface actually uses a struct from the hostdb subpackage,
-// which is a bit of a bad practice. The alternative would be to create a
-// different package to manage things like host entries and host announcements,
-// but I think it makes enough sense to define them in the same package that
-// also provides an example (and the primary) implementation of the hostdb
-// interface.
+type HostDB struct {
+	hostTree      *hostNode
+	activeHosts   map[string]*hostNode
+	inactiveHosts map[string]*components.HostEntry
+
+	dbLock sync.RWMutex
+}
+
+// New returns an empty HostDatabase.
+func New() (hdb *HostDB) {
+	hdb = &HostDB{
+		activeHosts:   make(map[string]*hostNode),
+		inactiveHosts: make(map[string]*components.HostEntry),
+	}
+	return
+}
+
+// Insert adds an entry to the hostdb.
+func (hdb *HostDB) Insert(entry components.HostEntry) error {
+	hdb.lock()
+	defer hdb.unlock()
+
+	_, exists := hdb.activeHosts[entry.ID]
+	if exists {
+		return errors.New("entry of given id already exists in host db")
+	}
+
+	if hdb.hostTree == nil {
+		hdb.hostTree = createNode(nil, &entry)
+		hdb.activeHosts[entry.ID] = hdb.hostTree
+	} else {
+		_, hostNode := hdb.hostTree.insert(&entry)
+		hdb.activeHosts[entry.ID] = hostNode
+	}
+	return nil
+}
+
+// Remove deletes an entry from the hostdb.
+func (hdb *HostDB) Remove(id string) error {
+	hdb.lock()
+	defer hdb.unlock()
+
+	node, exists := hdb.activeHosts[id]
+	if !exists {
+		_, exists := hdb.inactiveHosts[id]
+		if exists {
+			delete(hdb.inactiveHosts, id)
+			return nil
+		} else {
+			return errors.New("id not found in host database")
+		}
+	}
+	delete(hdb.activeHosts, id)
+	node.remove()
+
+	return nil
+}
+
+// Update throws a bunch of blocks at the hostdb to be integrated.
 //
-// Maybe though we can make the HostAnnouncement and the HostEntry their own
-// interfaces.
+// TODO: Check for repeat host announcements when parsing blocks.
+func (hdb *HostDB) Update(initialStateHeight consensus.BlockHeight, rewoundBlocks []consensus.Block, appliedBlocks []consensus.Block) (err error) {
+	hdb.lock()
+	defer hdb.unlock()
 
-type HostDB interface {
-	// Info returns an arbitrary byte slice presumably with information about
-	// the status of the hostdb. Info is not relevant to the sia package, but
-	// instead toa frontend.
-	Info() ([]byte, error)
+	// Remove hosts found in blocks that were rewound. Because the hostdb is
+	// like a stack, you can just pop the hosts and be certain that they are
+	// the same hosts.
+	for _, b := range rewoundBlocks {
+		var entries []components.HostEntry
+		entries, err = findHostAnnouncements(initialStateHeight, b)
+		if err != nil {
+			return
+		}
 
-	// Update gives the hostdb a set of blocks that have been applied and
-	// reversed.
-	Update(initialStateHeight consensus.BlockHeight, rewoundBlocks []consensus.Block, appliedBlocks []consensus.Block) error
+		for _, entry := range entries {
+			err = hdb.Remove(entry.ID)
+			if err != nil {
+				return
+			}
+		}
+	}
 
-	// Insert puts a host entry into the host database.
-	Insert(HostEntry) error
+	// Add hosts found in blocks that were applied.
+	for _, b := range appliedBlocks {
+		var entries []components.HostEntry
+		entries, err = findHostAnnouncements(initialStateHeight, b)
+		if err != nil {
+			return
+		}
 
-	// Remove pulls a host entry from the host database.
-	Remove(id string) error
+		for _, entry := range entries {
+			err = hdb.Insert(entry)
+			if err != nil {
+				return
+			}
+		}
+	}
 
-	// RandomHost pulls a host entry at random from the database, weighted
-	// according to whatever score is assigned the hosts.
-	RandomHost() (HostEntry, error)
+	return
+}
+
+// RandomHost pulls a random host from the hostdb weighted according to
+// whatever internal metrics exist within the hostdb.
+func (hdb *HostDB) RandomHost() (h components.HostEntry, err error) {
+	hdb.rLock()
+	defer hdb.rUnlock()
+	if len(hdb.activeHosts) == 0 {
+		err = errors.New("no hosts found")
+		return
+	}
+
+	// Get a random number between 0 and state.TotalWeight and then scroll
+	// through state.HostList until at least that much weight has been passed.
+	randInt, err := rand.Int(rand.Reader, big.NewInt(int64(hdb.hostTree.weight)))
+	if err != nil {
+		return
+	}
+	randWeight := consensus.Currency(randInt.Int64())
+	return hdb.hostTree.entryAtWeight(randWeight)
 }
