+++ sia/hostdb/hostentry.go
@@ -4,25 +4,8 @@ import (
 	"math"
 
 	"github.com/NebulousLabs/Sia/consensus"
-	"github.com/NebulousLabs/Sia/network"
+	"github.com/NebulousLabs/Sia/sia/components"
 )
-
-// the Host struct is kept in the client package because it's what the client
-// uses to weigh hosts and pick them out when storing files.
-type HostEntry struct {
-	ID          string
-	IPAddress   network.Address
-	MinFilesize uint64
-	MaxFilesize uint64
-	MinDuration consensus.BlockHeight
-	MaxDuration consensus.BlockHeight
-	Window      consensus.BlockHeight
-	Tolerance   uint64
-	Price       consensus.Currency
-	Burn        consensus.Currency
-	Freeze      consensus.Currency
-	CoinAddress consensus.CoinAddress
-}
 
 // host.Weight() determines the weight of a specific host, which is:
 //
@@ -38,10 +21,10 @@ type HostEntry struct {
 // We take the square of the price to heavily emphasize hosts that have a low
 // price. This is also a bit simplistic however, because we're not sure what
 // the host might be charging for bandwidth.
-func (h *HostEntry) Weight() consensus.Currency {
-	adjustedBurn := float64(h.Burn)
-	adjustedFreeze := float64(h.Freeze)
-	adjustedPrice := math.Sqrt(float64(h.Price))
+func entryWeight(entry *components.HostEntry) consensus.Currency {
+	adjustedBurn := float64(entry.Burn)
+	adjustedFreeze := float64(entry.Freeze)
+	adjustedPrice := math.Sqrt(float64(entry.Price))
 
 	weight := adjustedFreeze * adjustedBurn / adjustedPrice
 	return consensus.Currency(weight)
