+++ sia/hostdb/info.go
@@ -1,0 +1,5 @@
+package hostdb
+
+func (hdb *HostDB) Info() ([]byte, error) {
+	return nil, nil
+}
